plugins: Refactor DNS authenticator tests to avoid duplication
