+++ certbot/plugins/dns_test_common.py
@@ -1,0 +1,44 @@
+"""Base test class for DNS authenticators."""
+
+import mock
+import six
+
+from acme import challenges
+from acme import jose
+
+from certbot import achallenges
+
+from certbot.tests import acme_util
+from certbot.tests import util as test_util
+
+DOMAIN = 'example.com'
+KEY = jose.JWKRSA.load(test_util.load_vector("rsa512_key.pem"))
+
+
+class BaseAuthenticatorTest(object):
+    """
+    A base test class to reduce duplication between test code for DNS Authenticator Plugins.
+
+    Assumes:
+     * That subclasses also subclass unittest.TestCase
+     * That the authenticator is stored as self.auth
+    """
+
+    achall = achallenges.KeyAuthorizationAnnotatedChallenge(
+        challb=acme_util.DNS01, domain=DOMAIN, account_key=KEY)
+
+    def test_more_info(self):
+        # pylint: disable=no-member
+        self.assertTrue(isinstance(self.auth.more_info(), six.string_types))
+
+    def test_get_chall_pref(self):
+        # pylint: disable=no-member
+        self.assertEqual(self.auth.get_chall_pref(None), [challenges.DNS01])
+
+    def test_parser_arguments(self):
+        m = mock.MagicMock()
+
+        # pylint: disable=no-member
+        self.auth.add_parser_arguments(m)
+
+        m.assert_any_call('propagation-seconds', type=int, default=mock.ANY, help=mock.ANY)
